requirements.txt
fastapi[standard]>=0.116.1
httpx>=0.27.2
passlib[bcrypt]>=1.7.4
pre-commit>=4.3.0
prometheus-client>=0.20.0
prometheus-fastapi-instrumentator>=7.0.0
pydantic-settings>=2.10.1
pydantic>=2.4.2
python-dotenv>=1.0.0
python-jose[cryptography]>=3.5.0
redis[hiredis]>=5.0.8
uvicorn>=0.35.0

